Declare completion code as int
Matches libcurl (it declares them as enum) and avoids casting for
difference between 32/64bit

+++ CURLHandleSource/CURLTransfer+MultiSupport.h
@@ -38,7 +38,7 @@
 
  */
 
-- (void)completeWithCode:(NSInteger)code isMulti:(BOOL)isMultiCode;
+- (void)completeWithCode:(int)code isMulti:(BOOL)isMultiCode;
 
 /**
  Has the transfer completed?
